split actions and reducers
