+++ editor/source/actions/editor.js
@@ -1,0 +1,8 @@
+import { ACTION } from './index';
+
+export function setContent(content) {
+    return {
+        type: ACTION.SET_CONTENT,
+        content
+    }
+}
