+++ editor/source/app.js
@@ -6,7 +6,7 @@ import ReactDom from 'react-dom';
 import classNames from 'classnames';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { listAction } from './actions';
+import { menuAction } from './actions';
 
 import Editor from './components/editor';
 import List from './components/list';
@@ -17,20 +17,20 @@ import Toolbar from './components/toolbar';
 
 class App extends React.Component {
     componentWillMount() {
-        this.props.listAction.fetchList();
+        this.props.menuAction.fetchList();
     }
     render() {
         const menu = this.props.menu;
         const list = menu.get('list');
         const editor = this.props.editor;
-        const listAction = this.props.listAction;
+        const menuAction = this.props.menuAction;
         return (
             <div id="container">
-                <div id="left" onMouseLeave={listAction.hideList} onMouseOver={listAction.showList}>
+                <div id="left" onMouseLeave={menuAction.hideList} onMouseOver={menuAction.showList}>
                     <Menu menu={menu} />
                     <div id="files" className={classNames({hide: !menu.get('show')})}>
                         <Search />
-                        <List list={list} onOpenArticle={listAction.openArticle} />
+                        <List list={list} onOpenArticle={menuAction.openArticle} />
                     </div>
                 </div>
                 <div id="right">
@@ -47,6 +47,6 @@ export default connect(function(state) {
     return state;
 }, function(dispatch) {
     return {
-        listAction: bindActionCreators(listAction, dispatch)
+        menuAction: bindActionCreators(menuAction, dispatch)
     };
 })(App);
