+++ editor/source/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import menu from './menu';
+import editor from './editor';
+
+export default combineReducers({
+    menu,
+    editor
+});
